
\name{ExcessHeatIndices-package}
\alias{ExcessHeatIndices-package}
\alias{ExcessHeatIndices}
\docType{package}
\title{
Excess Heat Indices 
}
\description{
Excess Heat Indices for Human Health research
}
\details{
\tabular{ll}{
Package: \tab ExcessHeatIndices\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2017-09-01\cr
License: \tab GPL2\cr
}

}
\author{
ivanhanigan, James Goldie  
Maintainer: Who to complain to  ivan.hanigan@gmail.com 
}
\references{
Wilson, L. A., Morgan, G., Hanigan, I. C., Johnston, F., Abu-Rayya, H., Broome, R., … Jalaludin, B. (2013). The impact of heat on mortality and morbidity in the Greater Metropolitan Sydney Region: a case crossover analysis. Environmental Health : A Global Access Science Source, 12(1), 98. http://doi.org/10.1186/1476-069X-12-98
}

\keyword{ Heatwaves }
